vendor-qbr-analysis/requirements.txt
# データ処理
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.0

# 統計分析
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
pingouin>=0.5.3

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# レポート生成
reportlab>=4.0.0
jinja2>=3.1.2

# ユーティリティ
openpyxl>=3.1.0
python-dateutil>=2.8.2
tqdm>=4.66.0

# KDE、統計量計算に使用
scipy>=1.7.0  